enhancement(project): Translate Cookies Disclaimer to French
