+++ src/lib/frontend/components/CookiesDisclaimerModal.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
     import {wereCookiesAccepted} from '$lib/frontend/stores/localStorageStore/AcceptCookiesStore'
-    import {showCookiesDisclaimer} from "$lib/frontend/stores/localStorageStore/ShowCookiesDisclaimerStore";
+    import {showCookiesDisclaimer} from "$lib/frontend/stores/localStorageStore/ShowCookiesDisclaimerStore"
+    import t from 'sveltekit-i18n'
 
     wereCookiesAccepted.subscribe(value => {
         if (!value) {
@@ -15,34 +16,35 @@
 </script>
 
 <div class="fixed bottom-0 p-1 right-0 bg-green-50 shadow-md cookies-container block}">
-        <div class="border p-4 border-gray-600">
-            <p class="font-light text-xl">Your privacy, your choice!</p>
+    <div class="border p-4 border-gray-600">
+        <p class="font-light text-xl">{t('title')}</p>
 
-            <br>
+        <br>
 
-            <p class="font-thin">
-                We use cookies to enhance your experience on our website. Your experience may be hindered if you decline
-                our
-                use
-                of cookies.
-            </p>
+        <p class="font-thin">
+            We use cookies to enhance your experience on our website. Your experience may be hindered if you decline
+            our
+            use
+            of cookies.
+        </p>
 
-            <br>
+        <br>
 
-            <p class="font-thin">
-                By clicking accept you consent to our <a href="./">privacy policies.</a>
-            </p>
+        <p class="font-thin">
+            By clicking accept you consent to our <a href="./">privacy policies.</a>
+        </p>
 
-            <div class="flex justify-center mt-4">
-                <button data-testid="decline" class="btn variant-soft rounded-none mr-5" on:click={() => updateCookies(false)}>
-                    Decline
-                </button>
+        <div class="flex justify-center mt-4">
+            <button data-testid="decline" class="btn variant-soft rounded-none mr-5"
+                    on:click={() => updateCookies(false)}>
+                Decline
+            </button>
 
-                <button data-testid="accept" class="btn variant-filled rounded-none" on:click={() => updateCookies(true)}>
-                    Accept
-                </button>
+            <button data-testid="accept" class="btn variant-filled rounded-none" on:click={() => updateCookies(true)}>
+                Accept
+            </button>
         </div>
-            </div>
+    </div>
 </div>
 
 <style lang="css">
