--- src/lib/frontend/translations/en/home.json
@@ -1,4 +1,0 @@
-{
-  "title": "Welcome to home page",
-  "message": "A welcome message to static web site powered by  Sveltekit"
-}
