+++ src/translations/en/lib/frontend/components/cookies-disclaimer.json
@@ -1,0 +1,6 @@
+{
+  "title": "Your privacy, your choice!",
+  "body": "We use cookies to enhance your experience on our website. Your experience may be hindered if you decline\n            our\n            use\n            of cookies.",
+  "accept": "Accept",
+  "decline": "Decline"
+}
