+++ src/translations/en/routes/layout.json
@@ -1,0 +1,4 @@
+{
+  "title": "Welcome to home page",
+  "message": "A welcome message to static web site powered by Sveltekit"
+}
