+++ src/translations/fr/home.json
